add fix to update the product's quantity when a sell is done

+++ src/main/java/com/lcrtech/productregistration/dto/ItemDTO.java
@@ -14,6 +14,6 @@ public class ItemDTO {
     private Long id;
     private BigDecimal subTotal;
     private Integer quantity;
-    private Integer productId;
+    private Long productId;
     private String productName;
 }
